requirements.txt
Django==2.1.5
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
PyJWT==1.7.1
pytz==2018.9
